Moved Model class definition to it's own file
